pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "met-cu"
version = "0.2.1"
description = "Custom GPU kernels for every meteorological calculation. GPU-accelerated MetPy/metrust (CUDA + Metal)."
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["cuda", "metal", "gpu", "meteorology", "weather", "cupy", "apple-silicon"]
dependencies = [
    "numpy>=1.20",
]

[project.optional-dependencies]
gpu = ["cupy-cuda12x>=13.0"]
cuda = ["cupy-cuda12x>=13.0"]
metal = ["pyobjc-framework-Metal>=10.0; sys_platform == 'darwin'"]
plot = [
    "matplotlib>=3.10",
    "cartopy>=0.25",
    "Pillow>=10.0",
]
cpu = ["metrust>=0.3.8"]
test = [
    "metrust>=0.3.8",
    "pytest",
    "matplotlib>=3.10",
    "cartopy>=0.25",
    "Pillow>=10.0",
]

[tool.setuptools.packages.find]
where = ["python"]
